Add files via upload
In JavaScript, Boolean values are used to represent truth values, with true and false being the only two possible values. Boolean values are typically used in conditional statements and loops to control program flow.

Comparison operators are used in JavaScript to compare two values and return a Boolean result. The following comparison operators are available in JavaScript:

==: equal to (compares values for equality, but not necessarily data type)
===: strictly equal to (compares values and data types for equality)
!=: not equal to (compares values for inequality, but not necessarily data type)
!==: strictly not equal to (compares values and data types for inequality)
<: less than
>: greater than
<=: less than or equal to
>=: greater than or equal to